+++ apps/sanity/schemaTypes/index.ts
@@ -1,0 +1,3 @@
+import {activitySchema} from './activity'
+
+export const schemaTypes = [activitySchema]